agency-swarm-cursor-agent/presentation_agency/requirements.txt
agency-swarm==0.4.1
PyPDF2>=3.0.0
python-pptx>=0.6.21
matplotlib>=3.7.1
python-dotenv>=1.0.0 